Allow reallocating dynamic vertex, index and uniform buffers
Vertex, index and uniform buffers had a fixed size. With this change,
they can be reallocated when needed and grow to hold more data.

Because Vulkan is async, the existing buffers, their memory and
descriptor sets cannot be deleted or updated right away after allocating
new ones. Instead, the existing buffers need to be kept around for some
time until the device is done using them. That's the purpose of the
"to_be_deleted" fields added to some structures.

Dynamic buffer sizes grow with a factor of 1.5 in each reallocation, as
is common practice[1].

In addition, the common code in R_InitDynamicVertexBuffers,
R_InitDynamicIndexBuffers and R_InitDynamicUniformBuffers has been
abstracted to a single function, and all of them can be called again
when the buffers need to grow in size.

[1] https://stackoverflow.com/questions/2269063/buffer-growth-strategy