Fix dedicated server tic timing (#425)
The dedicated server limiter sys_ticrate interacts badly with host_maxfps. In the default configuration, every 50ms 3-4 server tics are run back to back and sent to the clients at once, which not only is wasteful but also breaks client interpolation.
This is a minimal fix, which effectively limits dedicated server tic rate to the minimum of both variables.
Also bump default dedicated server tic rate from 20 to 40 Hz.